// src/app/models/user.ts
export interface User {
  displayName: string;
  firstName: string;
  surname: string;
  emailAddress: string;
  type: string;
  userLocation: string;
  RFIDtoken: string;
  timeAttendance: boolean;
  wage: string;
  username: string;
  backOfficePassword: string;
  fourDigitPassword: string;
  frontOfficeFastLogin: boolean;
}